input to input legacy
