+++ .changeset/little-berries-rescue.md
@@ -1,0 +1,8 @@
+---
+"@salt-ds/lab": minor
+---
+
+**Breaking change**
+
+Rename `Input` to `InputLegacy`
+All API tokens with `--saltInput-` prefix changed to `--saltInputLegacy-` prefix
