+++ packages/lab/src/color-chooser/AlphaInputField.tsx
@@ -1,7 +1,7 @@
 import { clsx } from "clsx";
 import { useEffect, useState } from "react";
 import { makePrefixer } from "@salt-ds/core";
-import { Input } from "../input";
+import { InputLegacy as Input } from "../input-legacy";
 
 import { useWindow } from "@salt-ds/window";
 import { useComponentCssInjection } from "@salt-ds/styles";
