+++ packages/lab/src/combo-box/ComboBox.tsx
@@ -3,7 +3,10 @@ import { ForwardedRef, forwardRef, ReactElement, useCallback } from "react";
 
 import { DropdownBase, DropdownBaseProps } from "../dropdown";
 import { List, ListProps } from "../list";
-import { Input, InputProps } from "../input";
+import {
+  InputLegacy as Input,
+  InputLegacyProps as InputProps,
+} from "../input-legacy";
 
 import {
   CollectionItem,
