+++ packages/lab/src/input-legacy/StaticInputAdornment.tsx
