+++ packages/lab/src/input-legacy/useCursorOnFocus.ts
