--- packages/lab/src/input/Input.css
@@ -1,137 +1,0 @@
-/* Styles applied to the root element dependent on density */
-.salt-density-touch {
-  --input-adornment-height: calc(var(--salt-size-base) + var(--salt-size-unit) * 0.5);
-  --input-button-inset: 4px;
-}
-.salt-density-low {
-  --input-adornment-height: calc(var(--salt-size-base) + var(--salt-size-unit) * 2 / 3);
-  --input-button-inset: 4px;
-}
-.salt-density-medium {
-  --input-adornment-height: calc(var(--salt-size-base) + var(--salt-size-unit) * 0.5);
-  --input-button-inset: 2px;
-}
-.salt-density-high {
-  --input-adornment-height: calc(var(--salt-size-base) + var(--salt-size-unit));
-  --input-button-inset: 2px;
-}
-
-/* Style applied to the root element */
-.saltInput {
-  align-items: center;
-  background: var(--saltInput-background, none);
-  border: var(--saltInput-border, none);
-  border-radius: var(--saltInput-borderRadius, 0);
-  color: var(--saltInput-text-color, var(--salt-text-primary-foreground));
-  cursor: var(--saltInput-cursor, default);
-  display: inline-flex;
-  font-family: var(--saltInput-fontFamily, var(--salt-text-fontFamily));
-  font-size: var(--saltInput-fontSize, var(--salt-text-fontSize));
-  height: var(--saltInput-height, var(--salt-size-base));
-  line-height: var(--salt-text-lineHeight);
-  min-height: var(--saltInput-minHeight, var(--salt-size-base));
-  min-width: var(--saltInput-minWidth, 8em);
-  padding: var(--saltInput-padding, 0 var(--salt-size-unit));
-  position: var(--saltInput-position, relative);
-  width: 100%;
-}
-
-/* Reset in the next class */
-.saltInput-input:focus {
-  outline: none;
-}
-
-/* Pseudo-class applied to the root element when focused */
-.saltInput-focused {
-  outline-style: var(--saltInput-focused-outlineStyle, var(--salt-focused-outlineStyle));
-  outline-width: var(--saltInput-focused-outlineWidth, var(--salt-focused-outlineWidth));
-  outline-color: var(--saltInput-focused-outlineColor, var(--salt-focused-outlineColor));
-  outline-offset: var(--saltInput-focused-outlineOffset, var(--salt-focused-outlineOffset));
-}
-
-/* Style applied to selected input */
-.saltInput-input::selection {
-  background-color: var(--saltInput-highlight-color, var(--salt-text-background-selected));
-}
-
-/* Style applied to inner input component */
-.saltInput-input {
-  background: var(--saltInput-background, none);
-  border: none;
-  box-sizing: content-box;
-  color: inherit;
-  display: block;
-  flex: 1;
-  font: inherit;
-  letter-spacing: var(--saltInput-letterSpacing, 0);
-  margin: 0;
-  min-width: 0;
-  overflow: hidden;
-  padding: 0;
-  height: 100%;
-  width: 100%;
-}
-
-/* Style applied to input if `disabled={true}` */
-.saltInput-disabled .saltInput-input {
-  cursor: var(--salt-editable-cursor-disabled);
-  color: var(--saltInput-text-color-disabled, var(--salt-text-primary-foreground-disabled));
-}
-
-/* Style applied to adornment containers */
-.saltInput-suffixContainer,
-.saltInput-prefixContainer {
-  display: flex;
-  align-items: center;
-
-  height: var(--input-adornment-height);
-}
-
-/* Style applied to root element with start adornment */
-.saltInput-inputAdornedStart {
-  padding-left: var(--saltInput-adornedStart-padding, 0);
-}
-
-/* Style applied to inner input element with start adornment */
-.saltInput-inputAdornedStart .saltInput-input {
-  padding-left: var(--saltInput-adornedStart-input-padding, var(--salt-size-unit));
-}
-
-/* Style applied to root element with end adornment */
-.saltInput-inputAdornedEnd {
-  padding-right: var(--saltInput-adornedEnd-padding, 0);
-}
-
-/* Style applied to inner input element with end adornment */
-.saltInput-inputAdornedEnd .saltInput-input {
-  padding-right: var(--saltInput-adornedEnd-input-padding, var(--salt-size-unit));
-}
-
-/* Style applied to salt Button used within Input component adornments */
-.saltInput .saltInput-suffixContainer > .saltButton,
-.saltInput .saltInput-prefixContainer > .saltButton {
-  height: calc(var(--saltButton-height, var(--salt-size-base)) - (var(--input-button-inset) * 2));
-  margin: var(--input-button-inset);
-  padding: 0 calc(var(--salt-size-unit) - var(--input-button-inset));
-}
-
-/* Style applied if `textAlign={"left"}` */
-.saltInput-leftTextAlign .saltInput-input {
-  text-align: left;
-}
-
-/* Style applied if `textAlign={"center"}` */
-.saltInput-centerTextAlign .saltInput-input {
-  text-align: center;
-}
-
-/* Style applied if `textAlign={"right"}` */
-.saltInput-rightTextAlign .saltInput-input {
-  text-align: right;
-}
-
-/* Style applied if `inFormField={true}` */
-.saltInput-formField {
-  min-width: var(--saltFormFieldLegacy-input-minWidth, 0px);
-  width: 100%;
-}
