+++ packages/lab/stories/input-legacy/input-legacy.doc.stories.mdx
@@ -8,21 +8,26 @@ import {
 import { withFlexGap } from "docs/decorators/withFlexGap";
 
 <Meta
-  title="Documentation/Lab/Input"
+  title="Documentation/Lab/Input Legacy"
   component={Input}
   parameters={{
     viewMode: "docs",
   }}
 />
 
-# Input
+# Input Legacy
+
+<Banner status="warning">
+  This component has been deprecated. The new
+  [Input](?path=/docs/documentation-lab-input-next--page) is a current WIP.
+</Banner>
 
 The Input component provides an editable field in which users can enter arbitrary text and numeric values—with no specific format defined.
 
 ## Basic Input Component
 
 <Canvas>
-  <Story id="lab-input--feature-input" />
+  <Story id="lab-input-legacy--feature-input" />
 </Canvas>
 
 ## Read Only Input
@@ -30,7 +35,7 @@ The Input component provides an editable field in which users can enter arbitrar
 Read only mode switched on by providing the `readOnly` prop.
 
 <Canvas>
-  <Story id="lab-input--read-only" />
+  <Story id="lab-input-legacy--read-only" />
 </Canvas>
 
 ## Input with spellcheck
@@ -38,7 +43,7 @@ Read only mode switched on by providing the `readOnly` prop.
 Automatic spellchecking can be turned on by using `inputProps`.
 
 <Canvas>
-  <Story id="lab-input--spellcheck" />
+  <Story id="lab-input-legacy--spellcheck" />
 </Canvas>
 
 ## Input with adornments
@@ -46,7 +51,7 @@ Automatic spellchecking can be turned on by using `inputProps`.
 Icons and text can be used as adornments at the start or end of the component.
 
 <Canvas>
-  <Story id="lab-input--adornments" />
+  <Story id="lab-input-legacy--adornments" />
 </Canvas>
 
 ## Used in FormField
@@ -54,7 +59,7 @@ Icons and text can be used as adornments at the start or end of the component.
 Input is a component that would commonly be nested inside a FormField.
 
 <Canvas>
-  <Story id="lab-input--with-form-field" />
+  <Story id="lab-input-legacy--with-form-field" />
 </Canvas>
 
 # API
